fix: タスク定義比較の両辺を正規化し register cleanup を整理する
DDD_VIOLATIONS.md の #6 (High) と #11 に対応:

- AWS 生成フィールド除外ロジックを
  infrastructure/task-definition-normalizer.ts に共通化
- GitHub adapter の parseTaskDefinition でも正規化を適用
  (AWS export 形式の JSON が GitHub に置かれた場合に
  revision/taskDefinitionArn 等が常に差分として残る regression を回避)
- parseTaskDefinition で非 object 入力を検知して null を返すように強化
- registerTaskDefinition の cleanup から
  name 空の secret を黙って filter するロジックを削除
  (ユーザーの desired state を暗黙に改変するのを避ける。
  空配列の削除は AWS API 制約への適合として維持)
